ultimato

+++ Model/model.php
@@ -62,18 +62,18 @@
         public function deleteSong(){
             if(isset($_REQUEST['delete'])){
                 $this->connectToDB();
-                /*if(self::$mysqli->errno>0){
+                if(self::$mysqli->errno>0){
                     echo "deleteSong1";
                     return "Error";}
-                else{*/
+                else{
                     $song_id=$_REQUEST['delete'];
                     $del=self::$mysqli->query("DELETE FROM songs WHERE song_id='$song_id' AND favoriteBy IS NULL");
                     if(self::$mysqli->errno>0 || self::$mysqli->affected_rows!=1){
-                        echo "deleteSong1";
+                        echo "deleteSong2";
                         return "Error";}
                     else
                         return "Ok";
-                //}
+                }
             }
         } 
         //elenco artisti
